requirements.txt
# Core ML frameworks
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
triton==3.4.0

# Transformers ecosystem
transformers==4.56.2
accelerate==1.11.0
peft==0.17.1
trl==0.22.2
bitsandbytes==0.48.2

# Unsloth
unsloth==2025.11.2
unsloth_zoo==2025.11.3

# xformers
xformers==0.0.32.post2

# HuggingFace utilities
datasets==3.6.0
huggingface-hub==0.36.0
hf_transfer==0.1.9
tokenizers==0.22.1
safetensors==0.6.2

# Text processing
sentencepiece==0.2.1
protobuf==5.29.5

# Metrics
jiwer==4.0.0

# Utilities
einops==0.8.1
addict==2.4.0
easydict==1.13

# Image processing
pillow==11.3.0
opencv-python==4.12.0.88

# Scientific computing
numpy==2.0.2
scipy==1.16.3
scikit-learn==1.6.1

# Progress bars & CLI
tqdm==4.67.1
typer==0.20.0

# Optional (for logging/monitoring)
wandb==0.22.3
tensorboard==2.19.0